[Norm] norminette에 어긋나는 문법 수정
