+++ Makefile
@@ -6,7 +6,7 @@
 #    By: spark2 <spark2@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/01/15 15:37:35 by yerilee           #+#    #+#              #
-#    Updated: 2024/02/01 13:40:40 by spark2           ###   ########.fr        #
+#    Updated: 2024/02/04 02:39:30 by spark2           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -23,9 +23,9 @@ GNL = parsing/get_next_line/get_next_line.c parsing/get_next_line/get_next_line_
 
 SRCS = main.c \
 	execution/dda.c execution/draw.c execution/executing.c execution/key.c execution/move.c execution/utils.c \
-	parsing/check_map2.c parsing/check_overlap.c parsing/check.c parsing/direction.c parsing/ft_split.c parsing/init.c parsing/parsing.c parsing/read_map.c \
-	parsing/rgb.c parsing/utils.c \
-	libft/ft_atoi.c libft/ft_strcpy.c libft/ft_strncmp.c
+	parsing/check_map1.c parsing/check_map2.c parsing/check_overlap.c parsing/check.c parsing/direction.c \
+	parsing/ft_split.c parsing/init1.c parsing/init2.c parsing/parsing.c parsing/read_map.c \
+	parsing/rgb.c parsing/utils.c libft/ft_atoi.c libft/ft_strcpy.c libft/ft_strncmp.c
 OBJS = $(SRCS:.c=.o) $(GNL:.c=.o)
 
 all: $(NAME)
